+++ src/index.ts
@@ -1,7 +1,7 @@
 export { KafkaProducer } from './lib/kafkaHighLevelProducer';
 export { KafkaConsumer } from './lib/kafkaConsumer';
 
-import consumerMock = require('./mock/kafkaNconsumer.mock');
+export { KafkaConsumerMock } from './mock/kafkaConsumer.mock';
 import producerMock = require('./mock/kafkaNproducer.mock');
 
-export { consumerMock, producerMock };
+export { producerMock };