+++ src/index.ts
@@ -2,6 +2,4 @@ export { KafkaProducer } from './lib/kafkaHighLevelProducer';
 export { KafkaConsumer } from './lib/kafkaConsumer';
 
 export { KafkaConsumerMock } from './mock/kafkaConsumer.mock';
-import producerMock = require('./mock/kafkaNproducer.mock');
-
-export { producerMock };
+export { KafkaProducerMock } from './mock/kafkaProducer.mock';