+++ CHANGELOG.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- (consumer) added consumer interface
 ### Changed
 - (index) changed export style to typescript 
 - (producer) started migration of high level producer to typescript
 - (consumer) started migration of kafka consumer to typescript
+- (mock) started migration of mocked consumer to typescript
 ### Deprecated
 ### Fixed
 - (producer) fixed thrown error if kafka.producer.topicsPrefix was missing