+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - (consumer) added consumer interface
 ### Changed
 - (index) changed export style to typescript 
-- (producer) started migration of high level producer to typescript
-- (consumer) started migration of kafka consumer to typescript
-- (mock) started migration of mocked consumer to typescript
+- (producer) migration of high level producer to typescript
+- (consumer) migration of kafka consumer to typescript
+- (mock) migration of mocked consumer to typescript
+- (mock) migration of mocked producer to typescript
 ### Deprecated
 ### Fixed
 - (producer) fixed thrown error if kafka.producer.topicsPrefix was missing