+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - (producer) removed kafkaNProducer
 - (producer) removed poll() before disconnecting
+- (dependencies) removed node-config dependency
 ### Security
 
 ## [1.1.1] - 2020-01-07