+++ CHANGELOG.md
@@ -20,8 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - (producer) removed kafkaNProducer
 - (producer) removed poll() before disconnecting
-- (dependencies) removed node-config dependency
 - (dependencies) removed unused hirestime and uuid dependency
+- (dependencies) removed unused prom-client, s3pweb-logger and config dependency
 ### Security
 
 ## [1.1.1] - 2020-01-07